# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Orla Daemon API
  description: HTTP API for the Orla agent execution engine daemon (`orla serve`).
  version: 1.0.0

servers:
  - url: http://localhost:8081
    description: Default local server

paths:
  /api/v1/health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/execute:
    post:
      summary: Run inference
      description: |
        Execute a chat completion against a registered LLM backend. Supports both
        streaming (SSE) and non-streaming responses. Set `stream: true` for SSE.
      operationId: execute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequest"
      responses:
        "200":
          description: Success (non-streaming) or SSE stream start (streaming)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteResponse"
            text/event-stream:
              schema:
                type: string
                description: SSE stream with events content, thinking, done
        "400":
          description: Bad request (missing backend, invalid JSON, unsupported policy)
        "429":
          description: Rate limit exceeded (when rate_limit_rps is configured)
        "500":
          description: Internal error (backend not found, inference failed)

  /api/v1/backends:
    get:
      summary: List backends
      operationId: listBackends
      responses:
        "200":
          description: List of registered backend names
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListBackendsResponse"
    post:
      summary: Register backend
      description: "Register an LLM backend for inference. Types: openai, sglang."
      operationId: registerBackend
      responses:
        "200":
          description: Backend registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterBackendResponse"
        "400":
          description: Validation error (missing fields, invalid type)
        "429":
          description: Rate limit exceeded
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBackendRequest"

  /api/v1/backends/{name}:
    patch:
      summary: Update backend
      description: >
        Live-update mutable fields on a registered backend. Only supplied fields
        are changed; omitted fields are left as-is.
      operationId: updateBackend
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Name of the registered backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBackendRequest"
      responses:
        "200":
          description: Backend updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          description: Validation error
        "404":
          description: Backend not found

  /api/v1/workflow/complete:
    post:
      summary: Notify workflow complete
      description: Signals workflow completion for memory manager tracking.
      operationId: workflowComplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowCompleteRequest"
      responses:
        "200":
          description: Notification accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowCompleteResponse"
        "400":
          description: Missing workflow_id

  /metrics:
    get:
      summary: Prometheus metrics
      description: Prometheus-formatted metrics (orla_requests_total, orla_queue_wait_seconds, etc.)
      operationId: getMetrics
      responses:
        "200":
          description: Prometheus text format
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: healthy

    ExecuteRequest:
      type: object
      required: [backend]
      properties:
        backend:
          type: string
          description: Name of registered LLM backend
        stage_id:
          type: string
          description: Stage identifier for queueing (default "default")
        prompt:
          type: string
          description: User prompt (appended as user message if provided)
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        tools:
          type: array
          items:
            $ref: "#/components/schemas/McpTool"
        stream:
          type: boolean
          default: false
        max_tokens:
          type: integer
          nullable: true
        temperature:
          type: number
          nullable: true
        top_p:
          type: number
          nullable: true
        scheduling_policy:
          type: string
          enum: [fcfs, priority]
        request_scheduling_policy:
          type: string
          enum: [fifo, priority]
        scheduling_hints:
          type: object
          properties:
            priority:
              type: integer
        workflow_id:
          type: string
        cache_policy:
          type: string
        accuracy:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true
          description: >
            Requests cost-optimized backend selection. When set, the daemon picks
            the cheapest backend whose quality >= this value.
        accuracy_policy:
          type: string
          enum: [prefer, strict]
          default: prefer
          description: >
            Controls fallback when no backend meets the accuracy threshold.
            "prefer" (default) falls back to the cheapest costed backend;
            "strict" fails the request.

    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        tool_name:
          type: string
        tool_call_id:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCallWithId"

    ToolCallWithId:
      type: object
      required: [id]
      properties:
        id:
          type: string
        name:
          type: string
        arguments:
          type: object

    McpTool:
      type: object
      description: MCP tool definition (see go-sdk/mcp)

    ExecuteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        response:
          $ref: "#/components/schemas/Response"
        error:
          type: string

    Response:
      type: object
      properties:
        content:
          type: string
        thinking:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCallWithId"
        metrics:
          $ref: "#/components/schemas/ResponseMetrics"

    ResponseMetrics:
      type: object
      properties:
        ttft_ms:
          type: integer
          nullable: true
          description: Time to first token (ms). Streaming only; null for non-streaming.
        tpot_ms:
          type: integer
          nullable: true
          description: Time per output token (ms). Streaming only; null for non-streaming.
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        queue_wait_ms:
          type: integer
        scheduler_decision_ms:
          type: integer
        dispatch_ms:
          type: integer
        backend_latency_ms:
          type: integer
          nullable: true
          description: End-to-end backend latency (ms). Non-streaming only; null for streaming.
        estimated_cost_usd:
          type: number
          format: double
          nullable: true
          description: >
            Estimated cost in USD based on token counts and the backend's cost model.
            Null when the backend has no cost model configured.

    CostModel:
      type: object
      properties:
        input_cost_per_mtoken:
          type: number
          format: double
          description: USD per million input tokens
        output_cost_per_mtoken:
          type: number
          format: double
          description: USD per million output tokens

    RegisterBackendRequest:
      type: object
      required: [name, endpoint, type, model_id]
      properties:
        name:
          type: string
          description: Backend name (used as "backend" in execute)
        endpoint:
          type: string
          example: "http://vllm:8000/v1"
        type:
          type: string
          enum: [openai, sglang]
        model_id:
          type: string
          example: "openai:Qwen/Qwen3-4B-Instruct-2507"
        api_key_env_var:
          type: string
          description: Env var name for API key (openai-type)
        max_concurrency:
          type: integer
          default: 1
        queue_capacity:
          type: integer
          default: 4096
        cost_model:
          $ref: "#/components/schemas/CostModel"
        quality:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Relative capability score in [0.0, 1.0] for accuracy-based routing

    RegisterBackendResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        error:
          type: string

    ListBackendsResponse:
      type: object
      required: [backends]
      properties:
        backends:
          type: array
          items:
            type: string

    UpdateBackendRequest:
      type: object
      description: Partial update for a registered backend. Only supplied fields are changed.
      properties:
        cost_model:
          $ref: "#/components/schemas/CostModel"
        quality:
          type: number
          format: double
          minimum: 0
          maximum: 1
          description: Updated quality score
        max_concurrency:
          type: integer
          minimum: 1
          description: Updated max concurrency

    WorkflowCompleteRequest:
      type: object
      required: [workflow_id]
      properties:
        workflow_id:
          type: string
        backends:
          type: array
          items:
            type: string

    WorkflowCompleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        error:
          type: string